moved out snipe stuff from main to its own pkg
