+++ server/main.go
@@ -1,10 +1,11 @@
 package main
 
 import (
-	"io"
-	"text/template"
-  "net/http"
+  "io"
+  "text/template"
+	"net/http"
 
+	"beta/snipe"
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
 )
@@ -23,68 +24,13 @@ func NewTemplate() *Templates {
   }
 }
 
-type Device struct {
-  AssetTag int
-  SerialNumber string 
-  Name string
-  Company string
-  Title string
-  Notes string
-  StartDate string
-  CompletionDate string
-}
-
-func NewDevice(tag int, sn, name, company, title, notes, start, completion string) Device {
-  return Device {
-    AssetTag: tag,
-    SerialNumber: sn,
-    Name: name,
-    Company: company,
-    Title: title,
-    Notes: notes,
-    StartDate: start,
-    CompletionDate: completion,
-  }
-}
-
-type Devices = []Device
-type MaintenanceTableData struct {
-  Devices Devices 
-}
-
-func NewMaintenanceTableData() MaintenanceTableData {
-  return MaintenanceTableData {
-    Devices: []Device {
-      NewDevice(
-        1234,
-        "123asd0p",
-        "Test CB",
-        "Campus 001",
-        "Broken camera",
-        "camera is blurry", 
-        "2024-06-01",
-        ""),
-
-      NewDevice(
-        5678, 
-        "567hjk0p",
-        "Test Laptop",
-        "Campus 002",
-        "Broken bottom cover",
-        "right corner broke from hinge issue",
-        "2024-06-01",
-        ""),
-    },
-  }
-}
-
 type PageData struct {
-  MaintenanceTableData MaintenanceTableData
+  MaintenanceTableData snipe.MaintenanceTableData
 }
 
-func NewPageData() (PageData) {
+func NewPageData() PageData {
   return PageData {
-    MaintenanceTableData: NewMaintenanceTableData(),  
+    MaintenanceTableData: snipe.NewMaintenanceTableData(),  
   }
 }
