+++ server/snipe/go.mod
@@ -1,0 +1,3 @@
+module beta/snipe
+
+go 1.22.1
