+++ server/snipe/snipe.go
@@ -1,0 +1,5 @@
+package snipe
+
+func Sin() string {
+  return "this is a test from snipe"
+}
